public_html/backup/PROF_INSTAL/Mieszkancy/profinstal_web/requirements.txt

flask>=3.0.0
python-docx>=1.1.0
reportlab>=4.0.0
